requirements.txt
keystone-engine==0.9.2
capstone==5.0.1
cryptography==42.0.5
pytest==8.0.2